@@ -9,11 +9,13 @@
 		"email": "sindresorhus@gmail.com",
 		"url": "https://sindresorhus.com"
 	},
+	"type": "module",
+	"exports": "./index.js",
 	"engines": {
-		"node": ">=0.10.0"
+		"node": ">=12"
 	},
 	"scripts": {
-		"test": "ava"
+		"test": "xo && ava"
 	},
 	"files": [
 		"index.js"
@@ -33,9 +35,10 @@
 		"char"
 	],
 	"dependencies": {
-		"escape-string-regexp": "^1.0.2"
+		"escape-string-regexp": "^5.0.0"
 	},
 	"devDependencies": {
-		"ava": "*"
+		"ava": "^3.15.0",
+		"xo": "^0.39.1"
 	}
 }
