@@ -1,10 +1,11 @@
-'use strict';
-var escapeStringRegexp = require('escape-string-regexp');
+import escapeStringRegexp from 'escape-string-regexp';
 
-module.exports = function (str, target) {
-	if (typeof str !== 'string' || typeof target !== 'string') {
+export default function trimRepeated(string, target) {
+	if (typeof string !== 'string' || typeof target !== 'string') {
 		throw new TypeError('Expected a string');
 	}
 
-	return str.replace(new RegExp('(?:' + escapeStringRegexp(target) + '){2,}', 'g'), target);
-};
+	const regex = new RegExp(`(?:${escapeStringRegexp(target)}){2,}`, 'g');
+
+	return string.replace(regex, target);
+}
